convert saxswaxs.ZarrSetupWriter to saxswaxs.SetupResultsProcessor
